issue-24
Created a make offer component, will be added to each card that contains the item listing. Registered users will be able to 'make offers' and submit. The submitted offer should send to the owner of the listed item or pop up as part of their listings when they go to take a look at it.